separate icetray-dependent util functions from non-icetray dependent ones. Obviates need for icetray to make plots.
